+++ src/PlutoMathInput.jl
@@ -7,10 +7,6 @@ using JSON3
 
 export MathInput, MathDisplay
 
-# Stubs for Symbolics extension
-function to_symbolics end
-function from_symbolics end
-
 include("mathinput.jl")
 include("mathdisplay.jl")
 include("show.jl")